Move auth and question service routers to their on subdirectories
