+++ backend/auth-service/src/app.js
@@ -1,6 +1,6 @@
 import express from "express";
 import cors from "cors";
-import { authRoute } from "./routes.js";
+import { authRoutes } from "./routes/authRoutes.js";
 import { catchAllErrorHandler } from "./middlewares/errorHandlers.js";
 
 const app = express();
@@ -9,7 +9,7 @@ app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 app.use(cors());
 app.options("*", cors());
-app.use(authRoute);
+app.use(authRoutes);
 app.use(catchAllErrorHandler);
 
 export { app };
